+++ Puzzles/Puzzles/Puzzle5.cs
@@ -2,12 +2,12 @@
 
 namespace AoC2019.Puzzles;
 
-public class Puzzle5 : PuzzleBase<IEnumerable<int>, int, int>
+public class Puzzle5 : PuzzleBase<IEnumerable<long>, long, long>
 {
     protected override string Filename => "Input/puzzle-input-05.txt";
     protected override string PuzzleTitle => "--- Day 5: Sunny with a Chance of Asteroids ---";
     
-    public override int PartOne(IEnumerable<int> input)
+    public override long PartOne(IEnumerable<long> input)
     {
         var virtualMachine = new IntcodeComputer(new IntcodeConfig { InputValue = 1 }, input.ToArray());
         
@@ -16,7 +16,7 @@ public class Puzzle5 : PuzzleBase<IEnumerable<int>, int, int>
         return virtualMachine.GetOutput();
     }
 
-    public override int PartTwo(IEnumerable<int> input)
+    public override long PartTwo(IEnumerable<long> input)
     {
         var virtualMachine = new IntcodeComputer(new IntcodeConfig { InputValue = 5 }, input.ToArray());
         
@@ -25,8 +25,8 @@ public class Puzzle5 : PuzzleBase<IEnumerable<int>, int, int>
         return virtualMachine.GetOutput();
     }
 
-    public override IEnumerable<int> Preprocess(IPuzzleInput input, int part = 1)
+    public override IEnumerable<long> Preprocess(IPuzzleInput input, int part = 1)
     {
-        return input.GetFirstLine().Trim().Split(',').Select(int.Parse);
+        return input.GetFirstLine().Trim().Split(',').Select(long.Parse);
     }
 }
