+++ Puzzles/Util/Intcode/Instructions/OutputInstruction.cs
@@ -8,7 +8,7 @@ public class OutputInstruction : IOInstruction
     public OutputInstruction(Opcode opcode, (int m1, int m2, int m3) parameterModes) 
         : base(opcode, parameterModes) { }
     
-    public override void Run(int[] memory)
+    public override void Run(long[] memory)
     {
         ResultValue = GetParameterValue(memory, Parameter1);
     }
